ms-document-ai/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.6.0
sqlalchemy==2.0.36
pymysql==1.1.1
cryptography==43.0.3
alembic==1.13.3
python-jose[cryptography]==3.3.0
pytesseract==0.3.13
pdf2image==1.17.0
Pillow==11.0.0
opencv-python==4.10.0.84
numpy==2.1.3
spacy==3.8.2
requests==2.32.3
python-dotenv==1.0.1